app/requirements.txt
Flask
Werkzeug
joblib
pandas
numpy
google-api-python-client
google-auth
scikit-learn
dill